Check exit_code of compile commands in gnu packages

+++ utils.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 import tarfile
 
 import requests
@@ -9,6 +10,10 @@ def create_folder(name):
     os.mkdir(name)
 
 def download_file(url, output):
+    if os.path.exists(output):
+        print(f"File {output} already exists")
+        return
+
     with requests.get(url, stream=True) as r:
         total_size = int(r.headers["Content-Length"])
         with open(output, "wb") as f:
@@ -21,5 +26,14 @@ def download_file(url, output):
             print("")
 
 def extract_tarball(tarball, output_directory):
+    tarballfolder = Path(tarball).with_suffix('').stem
+    if os.path.exists(f"{output_directory}/{tarballfolder}"):
+        print(f"File already extracted {output_directory}/{tarballfolder}")
+        return
     with tarfile.open(tarball) as tar:
         tar.extractall(output_directory)
+
+def extract_debfile(debfile, output_directory):
+    os.system(f"ar vx {debfile} --output {output_directory}")
+    extract_tarball(f"{output_directory}/data.tar.xz", output_directory)
+
